update to support session timeout -
Add logic to send keep alive messages to the server. After 10 seconds of no keepalive messages, the session state is reset.

The idea is that both the client and server-side understand that they need to re-broker the keys if any errors occur, and we need a way to know that the client has "disconnected" given its over HTTP.